KangLin/tigervnc fd32c28515c83ec3db6b812f2ee8f5fbbcf29b33

+++ common/rdr/HexOutStream.cxx
@@ -20,10 +20,9 @@
 #ifdef HAVE_CONFIG_H
 #include <config.h>
 #endif
-
+#include <algorithm>
 #include <rdr/HexOutStream.h>
 #include <rfb/util.h>
-#include <algorithm>
 
 using namespace rdr;
 
@@ -40,7 +39,7 @@ bool HexOutStream::flushBuffer()
 {
   while (sentUpTo != ptr) {
     uint8_t* optr = out_stream.getptr(2);
-    size_t length = std::min(ptr-sentUpTo, (int64_t)out_stream.avail()/2);
+    size_t length = std::min((size_t)(ptr-sentUpTo), out_stream.avail()/2);
 
     for (size_t i=0; i<length; i++)
       rfb::binToHex(&sentUpTo[i], 1, (char*)&optr[i*2], 2);
@@ -63,4 +62,3 @@ void HexOutStream::cork(bool enable)
   BufferedOutStream::cork(enable);
   out_stream.cork(enable);
 }
-
